update registration and recover password commands

+++ src/main/java/com/epam/redkin/web/controller/command/user/LoginByTokenLinkCommand.java
@@ -7,6 +7,7 @@ import com.epam.redkin.web.controller.command.Command;
 import com.epam.redkin.web.listener.AppContext;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -21,16 +22,21 @@ public class LoginByTokenLinkCommand implements Command {
         UserService userService = AppContext.getInstance().getUserService();
         String redirectURL = PAGE_LOGIN;
         String token = request.getParameter(COOKIE_REMEMBER_USER_TOKEN);
-        try {
-            User user = userService.logInByToken(token);
-            request.getSession().setAttribute(USER_LOGIN, user.getEmail());
-            request.getSession().setAttribute(USER_ROLE, user.getRole().name());
-            request.getSession().setAttribute(USER_ID, user.getUserId());
-            userService.deleteRememberUserToken(user.getUserId());
-            redirectURL = PAGE_HOME;
-        } catch (ServiceException e) {
-            LOGGER.error("Incorrect data entered", e);
-            throw new ServiceException("Incorrect data entered");
+        if (StringUtils.isNoneBlank(token)) {
+            try {
+                User user = userService.logInByToken(token);
+                LOGGER.debug(user.toString());
+                request.getSession().setAttribute(EMAIL, user.getEmail());
+                request.getSession().setAttribute(USER_ROLE, user.getRole().name());
+                request.getSession().setAttribute(USER_ID, user.getUserId());
+                request.getSession().setAttribute(SESSION_USER, user);
+                System.out.println(user.getUserId());
+                userService.deleteRememberUserToken(user.getUserId());
+                redirectURL = PAGE_HOME;
+            } catch (ServiceException e) {
+                LOGGER.error("Incorrect data entered", e);
+                throw new ServiceException("Incorrect data entered");
+            }
         }
         LOGGER.info("done");
         return redirectURL;
